* Modifikasi API get daftar beberapa voucher yang belum di rating * Menambah logic total rating produk * Fixing recycler view android when rating clicked * testing 4 rating, 1 rating, not rated * Update Rating produk when pelanggan gift rating to voucher * Fixing layout item in produk * Menambah bonus poin setelah pelanggan memberikan rating

+++ app/src/main/java/id/telkomsel/merchant/model/ModelProduk.kt
@@ -21,7 +21,7 @@ data class ModelProduk(
     var view: Int = 0,
     var jumlah_rating: Int = 0,
     var total_rating: Int = 0,
-    var rating: Int = 0,
+    var rating: Float = 0.0f,
     var terjual: Int = 0,
     var created_by: Int = 0,
     var created_at: String = "",
